fish: update theme & prompt + update clean function

+++ fish/functions/clean.fish
@@ -3,17 +3,60 @@
 
 function clean -d "Clean specified directories - or if not specified the current working one"
 
-    set -l path
-    if test -z "$argv"
-        set path .
-    else
-        set path $argv
+  # Find pattern options
+  set -l pattern -name '#*#' -o -name '*~' -o -name '.*~' -o -name '*.pyc'
+  # Limit 'find' search depths to 1 by default
+  set -l recur   -maxdepth 1
+
+  # Other default value
+  set -l dry     false
+  set -l path    .
+  set -l to_rm
+
+  # Argument parser
+  set -l args  (getopt "adh" $argv)
+  [ $status -gt 0 ]; and return 1
+  set args (echo $args | sed 's/^\s//' | tr ' ' '\n')
+
+  set -l i 1
+  for arg in $args
+    switch $arg
+      case "-r"
+        set recur ""
+      case "-d"
+        set dry true
+      case "--"
+        set i (math $i + 1)
+        break
+      case "-h"
+        printf 'Usage: clean [-rdh] [directory ...]
+
+Options:
+    -r      Cleanup recursively: search file to cleanup recursively in path
+    -d      Dry run: print out found files without removing them
+    -h      Display help and exit
+'
+        return 0
     end
+    set i (math $i + 1)
+  end
 
-    set -l to_rm (find $path -maxdepth 1 -name '#*#' -o -name '*~' -o -name '.*~' -o -name '*.pyc')
+  if [ $i -le (count $args) ]
+    set path $args[$i..(count $args)]
+  end
 
+  # Cleanup logic
+  set to_rm (find $path $pattern)
+  if [ (count $to_rm) -eq 0 ]
+    echo "# $_: nothing to clean" 1>&2
+  else if [ "$dry" = "true" ]
+    echo "# $_: dry mode" 1>&2
+    echo $to_rm
+  else
+    echo "# $_: cleanup mode" 1>&2
     for i in $to_rm
-        rm -v "$i"
+      rm -v "$i"
     end
+  end
 
 end
